fix minor bugs

+++ oneflow/python/model/maskrcnn/rpn.py
@@ -267,8 +267,8 @@ class RPNLoss(object):
             # num_neg_anchors = flow.math.reduce_sum(gt_objectness_logits == flow.constant_scalar(value=0, dtype=flow.int8), name="rpn_num_neg_anchors")
             num_pos_anchors_mask = gt_objectness_logits == flow.constant_scalar(value=1, dtype=flow.int8)
             num_neg_anchors_mask = gt_objectness_logits == flow.constant_scalar(value=0, dtype=flow.int8)
-            num_pos_anchors = flow.math.reduce_sum(flow.cast(num_pos_anchors_mask, dtype=flow.int32))
-            num_neg_anchors = flow.math.reduce_sum(flow.cast(num_neg_anchors_mask, dtype=flow.int32))
+            num_pos_anchors = flow.math.reduce_sum(flow.cast(num_pos_anchors_mask, dtype=flow.float))
+            num_neg_anchors = flow.math.reduce_sum(flow.cast(num_neg_anchors_mask, dtype=flow.float))
             if self.cfg.MODEL.COLLECT_ACCURACY_METRICS:
                 accuracy.put_metrics(
                     {
